Add source
This is the initial commit containing the source code of the initial code repo transferred from private TFS into open source GitHub repo.

Note: There was an effort to insure the code is buildable, such as updating SDK and dependency versions to latest so the project is buildable with modern tools (i.e. can be built w/VS2019).

However, that is not a guarntee everything will work as expected. There will likely be small snafus that a clean up here or there is required.

Please open a pull request if you come across such a problem.